requirements.txt
peewee==3.15.1
python-telegram-bot==13.13
google-api-python-client==2.54.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
py3-validate-email==1.0.7
fastapi==0.79.0